Fix 76/update api endpoints (#80)
* Refactor adminEndpoints for consistency with
endpoint naming conventions

Updated the endpoint names to use kebab-case for better code readability
and maintainability.

* Refactor admin endpoint handling

Introduce centralized admin class endpoint and authentication headers
for section retrieval to streamline codebase. This enhances
maintainability and reusability.

* Add getAllTeachersEndpoint constant and use it in
SelectAdviserCombobox

Added a new constant for fetching all teachers' data and utilized it in
the SelectAdviserCombobox component to maintain consistency and
reusability. This change enhances code readability and simplifies future
updates related to the endpoint.