Server 99/teacher cant add a subject to a class (#103)
* Update(ClassroomModel)

* Update(teacherCo ntroller): Add addSubjectToClass function

* Update(teacherRoute): add /class/add-subject route

* Update(ClassroomModel): make the schedules an array

make the schedules an array so it can store mulitple days schedule for one subject

* Update(addSubjectToClass)

* Update(teacherController): Add updateSubjectClass

* Update(teacherController): Add deleteSubjectFromClass function

* Update getAssignedClasses function

* Update(updateStudentsInClass): Add adviser checking

* Refactor teacherController to include
authorization checks

Add authorization checks to teacherController endpoints to ensure that
only the adviser of a class can add, update, or delete subjects. This
enhances security and privacy by preventing unauthorized access.

* Refactor endpoint for updating students in class

Changed the endpoint route to make it more intuitive and descriptive by
updating the path from "/update-students-in-class" to
"/class/update-students". This change enhances clarity and consistency
in the API design. No known issues.