Fix 165/get specific class is fetching from localhost (#172)
* extract get specific class endpoint

* Refactor(ClassDetailsContext): use axiosInstance

* feat: Refactor API endpoint for fetching assigned
class details

Refactor the endpoint for fetching assigned class details to use the
centralized endpoint from the studentEndpoints configuration file,
improving maintainability and consistency across multiple components.
This change eliminates redundancy and reduces the potential for errors.
No known issues or significant consequences are expected.

* feat: Add endpoint for student assigned class
retrieval

Added a new endpoint for retrieving the details of the class assigned to
a student. The endpoint is now imported from the studentEndpoints
configuration file. This change enables easier management and
flexibility in handling student class details.