Fix 171/users should be able to change password (#179)
* feat: Add password drawer to Student sidebar

Add a password drawer to the Student sidebar for updating passwords,
enhancing user experience by providing a convenient way to manage
passwords. This change includes adding the necessary components and
functionality to the StudentSidebar component. The addition of this
feature aims to improve user accessibility and security. No associated
issues.

* Refactor(Drawer): add max width on md

* Refactor StudentSidebar to use
ChangePasswordDrawer

Instead of using separate Drawer components, StudentSidebar now utilizes
a specialized ChangePasswordDrawer component, streamlining the code and
improving maintainability. This change reduces redundancy and aligns
with best practices for code organization.

* Refactor ChangePasswordDrawer to manage password
change
- Added state and handlers for password fields
- Implemented change password logic using API calls
- Created new endpoint for updating student password

* Refactor ChangePasswordDrawer to accept userType
prop

Pass the userType prop to ChangePasswordDrawer component to dynamically
determine the endpoint for updating passwords based on the user's type.
This makes the component more flexible and adaptable to different user
types. Additionally, add new password endpoints for admin and teacher in
the config files.